sirius_cpi/requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
pandas==2.1.1
numpy==1.26.0
openpyxl==3.1.2
python-jose==3.3.0
passlib[bcrypt]